% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{Account$current}
\alias{Account$current}
\title{Get current account}
\value{
\link{Account} for the current workspace.
}
\description{
In an execution context, this is the workspace in which the execution
was run.

In the CLI context (when running \code{latch} commands) this is the
current setting of \verb{latch workspace}, which defaults to the user's personal
workspace.
}
\examples{
\dontrun{
Account$current()
}
}
\seealso{
Other Accounts APIs: 
\code{\link{Account}}
}
\concept{Accounts APIs}
