% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{Account}
\alias{Account}
\title{User or team workspace. Can be used to fetch related resources.}
\description{
\code{\link[=Account$current]{Account$current()}} is the typical way of creating an \code{Account}

If the current request signer (CLI user or execution context)
lacks permissions to fetch some information, the corresponding operations
will act as if the information does not exist. Update operations will usually
produce errors.
}
\examples{
Account$new("1")
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
acc <- Account$current()
acc$print()
acc$load()
acc$print()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
acc <- Account$current()
acc$get_display_name()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
acc <- Account$current()
acc$list_registry_projects()
\dontshow{\}) # examplesIf}

## ------------------------------------------------
## Method `Account$print`
## ------------------------------------------------

acc <- Account$current()
acc$print()
}
\seealso{
Other Accounts APIs: 
\code{\link{Account$current}}
}
\concept{Accounts APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{String. Unique identifier.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Account-new}{\code{Account$new()}}
\item \href{#method-Account-print}{\code{Account$print()}}
\item \href{#method-Account-load}{\code{Account$load()}}
\item \href{#method-Account-get_display_name}{\code{Account$get_display_name()}}
\item \href{#method-Account-list_registry_projects}{\code{Account$list_registry_projects()}}
\item \href{#method-Account-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-new"></a>}}
\if{latex}{\out{\hypertarget{method-Account-new}{}}}
\subsection{Method \code{new()}}{
Get the representation of a specific account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{String. The ID of the target account.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-print"></a>}}
\if{latex}{\out{\hypertarget{method-Account-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{acc <- Account$current()
acc$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-load"></a>}}
\if{latex}{\out{\hypertarget{method-Account-load}{}}}
\subsection{Method \code{load()}}{
(Re-)populate this account instance's cache.

Future calls to most getters will return immediately without
making a network request.

Always makes a network request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-get_display_name"></a>}}
\if{latex}{\out{\hypertarget{method-Account-get_display_name}{}}}
\subsection{Method \code{get_display_name()}}{
Get the display name of this account.

This is an opaque string that can contain any valid Unicode data.

Display names are \emph{not unique} and \emph{must never be used as identifiers}.
Use \code{id} instead.

For users, this is the user's full name as reported by the authenitcation
provider.

For teams, this is the team display name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$get_display_name(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the display name if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String value of the display name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-list_registry_projects"></a>}}
\if{latex}{\out{\hypertarget{method-Account-list_registry_projects}{}}}
\subsection{Method \code{list_registry_projects()}}{
List Registry projects owned by this workspace.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$list_registry_projects(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the project list if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of \link[=RegistryProject]{projects} owned by this workspace.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Account-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Account-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
