% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_header}
\alias{auth_header}
\title{Authorize API calls}
\usage{
auth_header()
}
\value{
String value of the \code{Authorization} HTTP header.
}
\description{
Valid in workflow executions and after signing in using Latch CLI
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
httr2::request("https://vacuole.latch.bio") \%>\%
  httr2::req_headers(Authorization = auth_header()) \%>\%
  httr2::req_dry_run()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
