% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql.R
\name{graphql_body}
\alias{graphql_body}
\title{Prepare a GraphQL request payload}
\usage{
graphql_body(req, query, variables = NULL)
}
\arguments{
\item{req}{\code{\link[httr2:request]{httr2::request()}} to modify}

\item{query}{String value of the GraphQL document}

\item{variables}{Named list of JSON values for the query parameters}
}
\value{
\code{\link[httr2:req_body]{httr2::req_body_json()}} with the payload
}
\description{
Prepare a GraphQL request payload
}
\examples{
\dontrun{
httr2::request("https://gql.example.com") \%>\%
  graphql_body(
    "query Hello($username: String!) { hello(name: $username) }",
    variables = list(username = "John")
  ) \%>\%
  httr2::req_dry_run()
}
}
\keyword{internal}
