% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql.R
\name{graphql_perform}
\alias{graphql_perform}
\title{Get and process the response to a prepared GraphQL request}
\usage{
graphql_perform(req)
}
\arguments{
\item{req}{\code{\link[httr2:request]{httr2::request()}} to perform}
}
\value{
JSON object containing the response
}
\description{
\enumerate{
\item Makes the HTTP request
\item Aborts on any GraphQL errors after pretty-printing them
\item Returns the \code{data} field of the response
}
}
\examples{
\dontrun{
latch_request(paste0("https://vacuole.", sdk_domain, "/graphql")) \%>\%
  graphql_body("query WhoAmI { accountInfoCurrent { id } }") \%>\%
  graphql_perform()
}
}
\keyword{internal}
