% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch-package.R
\docType{package}
\name{latch-package}
\alias{latch}
\alias{latch-package}
\title{latch: LatchBio platform SDK for the R programming language}
\description{
Provides programmatic access to the LatchBio platform (https://latch.bio).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/latchbio/latch-sdk-r}
  \item Report bugs at \url{https://github.com/latchbio/latch-sdk-r/issues}
}

}
\author{
\strong{Maintainer}: Max Smolin \email{max@latch.bio} (\href{https://orcid.org/0009-0007-7944-9239}{ORCID})

Other contributors:
\itemize{
  \item LatchBio, Inc. [copyright holder, funder]
}

}
\keyword{internal}
