% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{latch_request}
\alias{latch_request}
\title{Initialize an authorized API request}
\usage{
latch_request(url)
}
\arguments{
\item{url}{String. The request target.}
}
\value{
Pre-configured \code{\link[httr2:request]{httr2::request()}}.
}
\description{
Adds default options to \code{\link[httr2:request]{httr2::request()}}:
\itemize{
\item User agent
\item Retry config
\item Timeout
\item Authorization using \code{\link[=auth_header]{auth_header()}}
}
}
\examples{
\dontrun{
latch_request("https://vacuole.latch.bio") \%>\%
  httr2::req_dry_run()
}
}
\keyword{internal}
