% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldata.R
\name{LDataNode$from_id}
\alias{LDataNode$from_id}
\title{Construct an \link[=LDataNode]{Latch Data node object} using a node identifier.}
\value{
\link{LDataNode} referencing the provided ID.
}
\description{
Construct an \link[=LDataNode]{Latch Data node object} using a node identifier.
}
\examples{
LDataNode$from_id("1234")
}
\seealso{
Other Latch Data APIs: 
\code{\link{LDataNode}}
}
\concept{Latch Data APIs}
