% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldata.R
\name{LDataNode}
\alias{LDataNode}
\title{Remote latch node (file, directory, or special object).}
\description{
Remote latch node (file, directory, or special object).

Remote latch node (file, directory, or special object).
}
\seealso{
Other Latch Data APIs: 
\code{\link{LDataNode$from_id}}
}
\concept{Latch Data APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{remote_path}}{String. Path within the remote Latch Data storage.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LDataNode-new}{\code{LDataNode$new()}}
\item \href{#method-LDataNode-print}{\code{LDataNode$print()}}
\item \href{#method-LDataNode-clone}{\code{LDataNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LDataNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-LDataNode-new}{}}}
\subsection{Method \code{new()}}{
Get a Latch Data node by path.

The path is not validated in any way.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LDataNode$new(remote_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remote_path}}{The path of the target node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LDataNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-LDataNode-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LDataNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LDataNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LDataNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LDataNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
