% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_utils.R
\name{registry_to_r_literal}
\alias{registry_to_r_literal}
\title{Adapt Registry value into the corresponding R value.}
\usage{
registry_to_r_literal(registry_literal, registry_type)
}
\arguments{
\item{registry_literal}{JSON object containing the raw Registry literal value.}

\item{registry_type}{JSON object containing the Registry type definition.}
}
\value{
R equivalent of the registry literal.
}
\description{
Adapt Registry value into the corresponding R value.
}
\keyword{internal}
