% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_types.R
\name{RegistryColumn}
\alias{RegistryColumn}
\title{Column definition for a Registry table.}
\description{
\href{RegistryTable.html#method-RegistryTable-get_columns}{\code{RegistryTable$get_columns()}}
is the typical way to get a \code{RegistryColumn}.
}
\seealso{
Other Registry APIs: 
\code{\link{RegistryInvalidValue}},
\code{\link{RegistryProject}},
\code{\link{RegistryRecord}},
\code{\link{RegistryTable}}
}
\concept{Registry APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{String. Unique identifier within the table. Not globally unique.}

\item{\code{def}}{Default value for new records.}

\item{\code{upstream_type}}{Raw column type. Used to convert between R values and Registry values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegistryColumn-new}{\code{RegistryColumn$new()}}
\item \href{#method-RegistryColumn-print}{\code{RegistryColumn$print()}}
\item \href{#method-RegistryColumn-clone}{\code{RegistryColumn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryColumn-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryColumn-new}{}}}
\subsection{Method \code{new()}}{
Create a Registry column definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryColumn$new(key, def, upstream_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{String. The column key.}

\item{\code{def}}{The default value.}

\item{\code{upstream_type}}{The raw column type.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryColumn-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryColumn-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this column definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryColumn$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryColumn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryColumn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryColumn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
