% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_types.R
\name{RegistryInvalidValue}
\alias{RegistryInvalidValue}
\title{\link[=RegistryRecord]{Registry record} value that failed validation.}
\description{
\link[=RegistryRecord]{Registry record} value that failed validation.

\link[=RegistryRecord]{Registry record} value that failed validation.
}
\seealso{
Other Registry APIs: 
\code{\link{RegistryColumn}},
\code{\link{RegistryProject}},
\code{\link{RegistryRecord}},
\code{\link{RegistryTable}}
}
\concept{Registry APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{raw_value}}{String. User-provided string representation of the invalid value.

May be \code{""} (the empty string) if the value is missing but
the column is required.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegistryInvalidValue-new}{\code{RegistryInvalidValue$new()}}
\item \href{#method-RegistryInvalidValue-print}{\code{RegistryInvalidValue$print()}}
\item \href{#method-RegistryInvalidValue-clone}{\code{RegistryInvalidValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryInvalidValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryInvalidValue-new}{}}}
\subsection{Method \code{new()}}{
Create a Registry invalid value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryInvalidValue$new(raw_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_value}}{String. The underlying string value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryInvalidValue-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryInvalidValue-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryInvalidValue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryInvalidValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryInvalidValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryInvalidValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
