% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_project.R
\name{RegistryProject}
\alias{RegistryProject}
\title{Registry project (folder containing \link[=RegistryTable]{Registry tables}).}
\description{
\href{Account.html#method-Account-list_registry_projects}{\code{Account$list_registry_projects()}}
is the typical way to get a \code{RegistryProject}.
}
\seealso{
Other Registry APIs: 
\code{\link{RegistryColumn}},
\code{\link{RegistryInvalidValue}},
\code{\link{RegistryRecord}},
\code{\link{RegistryTable}}
}
\concept{Registry APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{String. Unique identifier.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegistryProject-new}{\code{RegistryProject$new()}}
\item \href{#method-RegistryProject-print}{\code{RegistryProject$print()}}
\item \href{#method-RegistryProject-load}{\code{RegistryProject$load()}}
\item \href{#method-RegistryProject-get_display_name}{\code{RegistryProject$get_display_name()}}
\item \href{#method-RegistryProject-list_tables}{\code{RegistryProject$list_tables()}}
\item \href{#method-RegistryProject-clone}{\code{RegistryProject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-new}{}}}
\subsection{Method \code{new()}}{
Get the representation of a specific project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{String. The ID of the target project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-load"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-load}{}}}
\subsection{Method \code{load()}}{
(Re-)populate this project instance's cache.

Future calls to most getters will return immediately without
making a network request.

Always makes a network request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-get_display_name"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-get_display_name}{}}}
\subsection{Method \code{get_display_name()}}{
Get the display name of this project.

This is an opaque string that can contain any valid Unicode data.

Display names are \emph{not unique} and \emph{must never be used as identifiers}.
Use \code{id} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$get_display_name(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the display name if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String value of the display name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
List tables contained in this project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$list_tables(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the table list if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of \link[=RegistryTable]{tables} contained in this project.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryProject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryProject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryProject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
