% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_record.R
\name{RegistryRecord}
\alias{RegistryRecord}
\title{Registry record.}
\description{
\href{RegistryTable.html#method-RegistryTable-list_records}{\code{RegistryTable$list_records()}}
is the typical way to get a \code{RegistryRecord}.
}
\seealso{
Other Registry APIs: 
\code{\link{RegistryColumn}},
\code{\link{RegistryInvalidValue}},
\code{\link{RegistryProject}},
\code{\link{RegistryTable}}
}
\concept{Registry APIs}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{String. Unique identifier.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegistryRecord-new}{\code{RegistryRecord$new()}}
\item \href{#method-RegistryRecord-print}{\code{RegistryRecord$print()}}
\item \href{#method-RegistryRecord-load}{\code{RegistryRecord$load()}}
\item \href{#method-RegistryRecord-get_name}{\code{RegistryRecord$get_name()}}
\item \href{#method-RegistryRecord-get_columns}{\code{RegistryRecord$get_columns()}}
\item \href{#method-RegistryRecord-get_values}{\code{RegistryRecord$get_values()}}
\item \href{#method-RegistryRecord-clone}{\code{RegistryRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-new}{}}}
\subsection{Method \code{new()}}{
Get the representation of a specific record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{String. The ID of the target record.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable representation of this record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-load"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-load}{}}}
\subsection{Method \code{load()}}{
(Re-)populate this record instance's cache.

Future calls to most getters will return immediately without
making a network request.

Always makes a network request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the name of this record.

Names are unique within a table. Names are not globally unique.
Use \code{id} if a globally unique identifier is required.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$get_name(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the name if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String value of the name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-get_columns"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-get_columns}{}}}
\subsection{Method \code{get_columns()}}{
Get the columns of this record's table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$get_columns(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the columns if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list mapping between column keys and the
corresponding \link[=RegistryColumn]{columns}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-get_values}{}}}
\subsection{Method \code{get_values()}}{
Get this record's values.

The resulting dictionary is shared between all calls to this
instance's \code{get_values}.
Make deep copies if independent mutation is desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$get_values(load_if_missing = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{load_if_missing}}{Logical.\cr
If \code{TRUE}, \code{\link[=load]{load()}} the values if not in cache.
If \code{FALSE}, return \code{NULL} if not in cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list mapping between column keys and the corresponding values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegistryRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegistryRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegistryRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
