% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{vacuole_gql_perform}
\alias{vacuole_gql_perform}
\title{Make a request to Latch's internal Vacuole GraphQL API}
\usage{
vacuole_gql_perform(...)
}
\value{
JSON object containing the response
}
\description{
This API is intended for internal use only:
\itemize{
\item We will provide no documentation or support
\item There are no backwards compatibility guarantees
}
}
\examples{
\dontrun{
vacuole_gql_perform("query WhoAmI { accountInfoCurrent { id } }")
}
}
\keyword{internal}
